requirements.txt
beautifulsoup4==4.12.3
openai==1.23.2
python-dotenv==1.0.1
pandas==2.2.2
PySide6==6.7.0
sqlalchemy==2.0.29